refactor(home): extract components into seperate ones
